% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.hedgedrf}
\alias{predict.hedgedrf}
\title{hedgedrf prediction}
\usage{
\method{predict}{hedgedrf}(object, data, ...)
}
\arguments{
\item{object}{hedgedrf \code{hedgedrf} object.}

\item{data}{data New test data of class \code{data.frame} or
\code{gwaa.data} (GenABEL).}

\item{...}{Additional arguments to pass to the \code{predict.ranger}
function.}
}
\description{
hedgedrf prediction
}
